requirements.txt
praw
psycopg2-binary
python-dotenv
flask
PyYAML
gunicorn